Trato de armar la red neuronal. Dejo planteado como se llamaria para el ejercicio 3

+++ src/main/java/practico4/Util.java
@@ -7,7 +7,7 @@ import java.util.Random;
  */
 public class Util {
 
-    public static double randInt(double min, double max, Random rand) {
+    public static double randDouble(double min, double max, Random rand) {
 
         double result = min + (rand.nextDouble() * (max - min));
